streamlit/requirements.txt
streamlit
psycopg
psycopg-binary
pandas
google-genai